extracted searchbar component
